+++ config/variables.config.ts
@@ -1,0 +1,2 @@
+export const EXT_URL =
+  'https://chromewebstore.google.com/detail/media-flow-seek/phhigkiikolopghmahejjlojejpocagg?authuser=0&hl=en';